refactor: fix imports to be compatible with node es modules
