+++ packages/avif/codec/package.json
@@ -2,5 +2,6 @@
   "scripts": {
     "build": "../../../tools/build-cpp.sh && npm run patch-pre-script",
     "patch-pre-script": "cat pre.js >> enc/avif_enc.js && cat pre.js >> dec/avif_dec.js"
-  }
+  },
+  "type": "module"
 }
