+++ packages/jxl/codec/package.json
@@ -2,5 +2,6 @@
   "name": "jxl",
   "scripts": {
     "build": "../../../tools/build-cpp.sh"
-  }
+  },
+  "type": "module"
 }
