+++ packages/jxl/index.ts
@@ -1,3 +1,3 @@
-export { default as encode } from './encode';
-export { default as decode } from './decode';
-export { EncodeOptions } from './meta';
+export { default as encode } from './encode.js';
+export { default as decode } from './decode.js';
+export { EncodeOptions } from './meta.js';
