+++ packages/oxipng/codec/package.json
@@ -2,5 +2,6 @@
   "name": "oxipng",
   "scripts": {
     "build": "./build.sh"
-  }
+  },
+  "type": "module"
 }
