+++ packages/oxipng/index.ts
@@ -1,1 +1,1 @@
-export { default as optimise } from './optimise';
+export { default as optimise } from './optimise.js';
