+++ packages/oxipng/optimise.ts
@@ -16,8 +16,8 @@
  * - Use mutli-threading only in a Worker context
  * - Use multi-threading only if the browser has hardware concurrency > 1
  */
-import type { InitInput } from './codec/pkg/squoosh_oxipng';
-import { defaultOptions, OptimiseOptions } from './meta';
+import type { InitInput } from './codec/pkg/squoosh_oxipng.js';
+import { defaultOptions, OptimiseOptions } from './meta.js';
 import { threads } from 'wasm-feature-detect';
 
 async function initMT(moduleOrPath?: InitInput) {
@@ -25,7 +25,7 @@ async function initMT(moduleOrPath?: InitInput) {
     default: init,
     initThreadPool,
     optimise,
-  } = await import('./codec/pkg-parallel/squoosh_oxipng');
+  } = await import('./codec/pkg-parallel/squoosh_oxipng.js');
   await init(moduleOrPath);
   await initThreadPool(globalThis.navigator.hardwareConcurrency);
   return optimise;
@@ -33,7 +33,7 @@ async function initMT(moduleOrPath?: InitInput) {
 
 async function initST(moduleOrPath?: InitInput) {
   const { default: init, optimise } = await import(
-    './codec/pkg/squoosh_oxipng'
+    './codec/pkg/squoosh_oxipng.js'
   );
   await init(moduleOrPath);
   return optimise;
