+++ packages/png/encode.ts
@@ -16,8 +16,11 @@
  * and modified it to encode PNG images and also optimise them.
  */
 
-import type { InitInput, InitOutput as PngModule } from './codec/squoosh_png';
-import initPngModule, { encode as pngEncode } from './codec/squoosh_png';
+import type {
+  InitInput,
+  InitOutput as PngModule,
+} from './codec/squoosh_png.js';
+import initPngModule, { encode as pngEncode } from './codec/squoosh_png.js';
 
 let pngModule: Promise<PngModule>;
