+++ packages/resize/lib/resize/package.json
@@ -11,5 +11,6 @@
   ],
   "module": "squoosh_resize.js",
   "types": "squoosh_resize.d.ts",
-  "sideEffects": false
+  "sideEffects": false,
+  "type": "module"
 }
