+++ packages/webp/codec/package.json
@@ -1,5 +1,6 @@
 {
   "scripts": {
     "build": "../../../tools/build-cpp.sh"
-  }
+  },
+  "type": "module"
 }
