+++ packages/webp/decode.ts
@@ -15,10 +15,10 @@
  * Notice: I (Jamie Sinclair) have modified this file to accept an ArrayBuffer instead of typed array
  * and manually allow instantiation of the Wasm Module.
  */
-import type { WebPModule } from './codec/dec/webp_dec';
+import type { WebPModule } from './codec/dec/webp_dec.js';
 
-import webp_dec from './codec/dec/webp_dec';
-import { initEmscriptenModule } from './utils';
+import webp_dec from './codec/dec/webp_dec.js';
+import { initEmscriptenModule } from './utils.js';
 
 let emscriptenModule: Promise<WebPModule>;
