+++ packages/webp/index.ts
@@ -1,2 +1,2 @@
-export { default as encode } from './encode';
-export { default as decode } from './decode';
+export { default as encode } from './encode.js';
+export { default as decode } from './decode.js';
